rename to JsbBridgeWrapper
* add lisence
